frameworks/google-adk/llm-fact-check-agent-demo/pyproject.toml
[project]
name = "llm-fact-check-agent-demo"
version = "0.1.0"
description = "Google ADK LLM Fact-Checking Agent Demo - Automated fact verification and correction system"
readme = "README.md"
authors = [
    { name = "robert-moyai", email = "robert@moyai.ai" }
]
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.0.0",
    "google-cloud-aiplatform[adk,agent-engines]>=1.93.0",
    "google-genai>=1.9.0",
    "pydantic>=2.10.0",
]

[project.scripts]
llm-fact-check-agent-demo = "src.runner:main"

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
markers = [
    "integration: mark test as an integration test",
]
testpaths = ["tests"]
asyncio_mode = "auto"
